% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailles_effet.R
\name{tailles_effet}
\alias{tailles_effet}
\title{Calcul des tailles d'effet partielles (η², ω², ε²)}
\usage{
tailles_effet(F_val, df_effect, df_error)
}
\arguments{
\item{F_val}{Valeur F obtenue dans l'output de l'ANOVA (F.ratio)}

\item{df_effect}{Degrés de liberté de l'effet (df1)}

\item{df_error}{Degrés de liberté de l'erreur (df2)}
}
\value{
Un data.frame contenant les trois tailles d'effet partielles
}
\description{
Calcule les trois tailles d'effet partielles à partir des valeurs
F, df_effect et df_error issues d'une ANOVA factorielle ou d'effets simples.

Formules :
η²partiel = (F * df_effect) / (F * df_effect + df_error)
ω²partiel = (F * df_effect - df_effect) / (F * df_effect + df_error + 1)
ε²partiel = (F * df_effect - df_effect) / (F * df_effect + df_error)
}
\examples{
tailles_effet(F_val = 2.41, df_effect = 3, df_error = 318)

}
